Move identity specific modules to different pallet to avoid circular dependency for fiat filter pallet. Runtime benchmarks maybe broken
