+++ pallets/core_mods/Cargo.toml
@@ -1,0 +1,89 @@
+[package]
+name = "core_mods"
+version = "0.1.0"
+authors = ["Dock.io"]
+edition = "2018"
+license = 'Apache-2.0'
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
+
+[dependencies.serde]
+features = ['derive']
+optional = true
+version = '1.0.119'
+
+[dependencies.serde-big-array]
+optional = true
+version = "0.3.0"
+
+[dependencies.codec]
+default-features = false
+features = ['derive']
+package = 'parity-scale-codec'
+version = '1.3.6'
+
+[dependencies.sp-core]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.sp-std]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.sp-runtime]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.sp-io]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.frame-system]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.frame-support]
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dependencies.frame-benchmarking]
+optional = true
+default-features = false
+version = '2.0.1'
+git = "https://github.com/paritytech/substrate.git"
+branch = "frontier"
+
+[dev-dependencies]
+rand = "0.7.3"
+
+[features]
+default = ['std']
+std = [
+    'codec/std',
+    'frame-support/std',
+    'frame-benchmarking/std',
+    'serde',
+    'serde-big-array',
+    'sp-core/std',
+    'sp-io/std',
+    'sp-runtime/std',
+    'sp-std/std',
+    'frame-system/std',
+]
+
+runtime-benchmarks = [
+    'frame-benchmarking',
+]
